test(sbs3): run metrics tests in shared context
Fold the two metrics integration tests into one shared, cached Spring
context provided by a new AbstractMetricsSharedContextTest, so the module
loads a single metrics context instead of two.

- AbstractMetricsSharedContextTest: pin Main.class with RANDOM_PORT, import
  the metrics beans, the actuator MetricsEndpoint and JwtAuthTestConfiguration,
  and own the shared metric fixtures in a nested MetricsTestSupportConfiguration
  (the superset MeterBinder: six counters plus the for.value.transformations
  gauge). The parent never references its subclasses.
- AxelixMetricsEndpointTest / DefaultServiceMetricsGroupsFeedAssemblerTest:
  extend the shared base and drop their own @SpringBootTest/@Import and inner
  @TestConfiguration.

spring-test-profiler now reports both metrics test classes on a single cached
context entry.
